Move types to types file - RSC-1081
