+++ lib/src/components/NxFieldset/types.ts
@@ -23,3 +23,29 @@ export const propTypes: PropTypes.ValidationMap<Props> = {
   validationErrors: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.string.isRequired), PropTypes.string]),
   isPristine: PropTypes.bool
 };
+
+// The following are for the state helpers
+export type RadioValidator = (value: string | null) => ValidationErrors;
+export type CheckboxValidator = (values: string[]) => ValidationErrors;
+export type RadioSetter = (v: string | null) => void;
+export type CheckboxState = [boolean, () => void];
+export type CheckboxInitValues<K extends string | number> = Record<K, boolean>;
+export type CheckboxStates<K extends string | number> = Record<K, CheckboxState>;
+
+export interface RadioStateProps {
+  value: string | null;
+  isPristine: boolean;
+  validationErrors: ValidationErrors;
+}
+
+export interface CheckboxStateProps {
+  values: string[];
+  isPristine: boolean;
+  validationErrors: ValidationErrors;
+}
+
+export interface CheckboxGroupHookReturnValue<K extends string | number> {
+  states: CheckboxStates<K>;
+  isPristine: boolean;
+  validationErrors: ValidationErrors;
+}
